puhs

+++ database/migrations/2025_10_03_085812_create_product_recipes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained('product')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('ingredient_master_id')->constrained('ingredient_masters')->onUpdate('cascade')->onDelete('cascade');
-            $table->decimal('qty', 10, 2);
+            $table->double('qty');
             $table->string('created_by')->default('');
             $table->string('updated_by')->default('');
             $table->string('deleted_by')->default('');
